requirements.txt
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
dj-database-url==3.0.1
Django==5.2.4
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
Markdown==3.8.2
packaging==25.0
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
